% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplyColumn.R
\name{multiplyColumn}
\alias{multiplyColumn}
\title{mutates a numeric column by multiplying its
entries by a given numeric factor.}
\usage{
multiplyColumn(df, colname, fac)
}
\arguments{
\item{df}{short for data frame, refers to the input data frame.}

\item{colname}{short for column name, refers to the name of
the column that you would like to mutate.}

\item{fac}{short for factor, refers to the input numeric factor.}
}
\value{
the data frame with the new column as well as a
histogram of the new column.
}
\description{
This function takes a data frame, a column name
(whose entries are numeric), and a numeric value. It creates
a new column that multiplies the entries of the specified
column with the input factor and returns the data frame with
this new column, as well as a histogram of the new column.
}
\examples{
df = data.frame(matrix(rnorm(4),
                ncol=2, nrow=2,
                dimnames=list(NULL, c("x1", "x2"))))
multiplyColumn(df, "x1", 0.5)
multiplyColumn(df, "x2", 3)
}
